% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semVerClass.R
\docType{class}
\name{semVer}
\alias{semVer}
\title{semVer is a class to create an object for working with Semantic versioning.}
\format{An \code{\link{R6Class}} generator object}
\description{
semVer is a class to create an object for working with Semantic versioning.

semVer is a class to create an object for working with Semantic versioning.
}
\keyword{data}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{semVer$new()}}
\item \href{#method-clone}{\code{semVer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{semVer$new(Version = "", SemVerVersion = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{semVer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
